Rename example app libraries to common format
