+++ example/lib/Pages/palette_hue_picker_page.dart
