+++ example/lib/Pages/palette_picker_page.dart
