+++ example/lib/Pages/palette_saturation_picker_page.dart
