+++ example/lib/Pages/palette_value_picker_page.dart
