+++ example/lib/Pages/slider_picker_page.dart
