+++ example/lib/Pages/swatches_picker_page.dart
