+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import "package:flutter/cupertino.dart";
 import "package:flutter/material.dart";
 import "package:flutter/widgets.dart";
-import "MainPage.dart";
+import 'main_page.dart';
 
 void main() async {
   runApp(MyApp());
